micro-service-auth/src/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Aleks-Ti <mti61ro@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
grpcio = "^1.60.1"
grpcio-tools = "^1.60.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
